+++ src/index.ts
@@ -3,6 +3,7 @@
 
 export * from "./Api";
 export * from "./ApiTypes";
+export * from "./ApiConstants";
 
 export {isRequestError} from "./ApiUtils";